Add status field in the TaaS API
In the current implementation, when the TaaS API (tap-service,
tap-flow) call completes, the status of result is returned. [1]
In order to adopt an asynchronous model, the 'status' field is
needed in the API.
The following code adds the status field in the API.
In the first step, the status is set to 'ACTIVE' by default.

[1]: http://lists.openstack.org/pipermail/openstack-dev/2016-March/090088.html

Change-Id: Ib8e4ad15e5b4272ac7f8800f72d7f2c003db798e