Disable VLAN id checks in create_tap_flow and delete_tap_flow.
In order to satisfy the requirement that port MAC addresses are
unique only within a Neutron network, the ingress side flow for
mirroring traffic needs to match both the VLAN id of the network
and the MAC address of the port. However, the current implementation
of this logic doesn't work correctly because of the way in which OVS
handles VLANs and Neutron's use of the NORMAL action in br-int.

As a temporary workaround, the VLAN related check is being disabled.
The broad/multi-cast ingress flow is also being disabled because it
too relies on a VLAN id check.

When a reliable way to implement the VLAN id check is available
this logic will be restored.

Change-Id: I4a6c518baf568c5638eafbbae7495a66c113e515
Depends-On: Ie6b3811e41a94721679c9178cdd5119bdad8208d
Related-Bug: 1529595