tempest: Switch to local copy of tempset/scenario/manager.py
The manager.py file was literally copied from
tempest as of today. [1]

tempest/scenario/manager.py is not a part of tempest stable
interface for plugins. [2]

This was suggested on openstack-dev as a preparation for
the coming refactoring in tempest. [3]

As tempest is branchless, this change needs to be backported
to all relevant stable branches.

[1] d184d619d7adc6ab60055884515055a7f3cf7770
[2] https://docs.openstack.org/developer/tempest/plugin.html#stable-tempest-apis-plugins-may-use
[3] http://lists.openstack.org/pipermail/openstack-dev/2017-February/112938.html

Change-Id: I0797e2948e992fd3399e6c3683745610d0eaadff